move typegraphql folder
