+++ server/.gitignore
@@ -13,6 +13,7 @@ node_modules/
 
 # Build folder
 dist/
+src/generated/
 
 # next.js
 .next/
