+++ server/tsconfig.json
@@ -52,6 +52,6 @@
     "./src/**/*.tsx"
   ],
   "exclude": [
-    "node_modules"
+    "./node_modules",
   ]
 }
